[XLA:GPU] Split command buffer VA remapping into its own allocator class
Separate the command buffer VA remapping implementation from
GpuExecutableBufferAllocator into a new GpuExecutableVaRemapAllocator
subclass in gpu_executable_va_remap_allocator.{h,cc}:

- GpuExecutableBufferAllocator becomes the base class implementing the
  ALWAYS_UPDATE update mode (constants-only persistent allocations) and
  the runtime fallback behavior for the VA remapping modes. Its
  ExecutionScope is now polymorphic and exposes protected hooks
  (PrepareReservation, ResolveParameterBuffer, AllocateTransientBuffer,
  OnCopyProtectedOutput) at the points where an allocation-address
  policy can intervene.
- GpuExecutableVaRemapAllocator implements SKIP_TEMP and SKIP_PROFILED:
  the per-executor Remapping state, VA reservation layout, parameter
  Map()/escaping-buffer aliasing, step-alias release, and the
  SKIP_PROFILED profile state machine all move to the new file.
- A GpuExecutableBufferAllocator::Create factory picks the class from
  the update mode; CreateExecutionScope now returns
  std::unique_ptr<ExecutionScope>.

No functional change intended.
