+++ xla/service/gpu/BUILD
@@ -695,21 +695,23 @@ bool_flag(
 
 cc_library(
     name = "gpu_executable_buffer_allocator",
-    srcs = ["gpu_executable_buffer_allocator.cc"],
-    hdrs = ["gpu_executable_buffer_allocator.h"],
+    srcs = [
+        "gpu_executable_buffer_allocator.cc",
+        "gpu_executable_va_remap_allocator.cc",
+    ],
+    hdrs = [
+        "gpu_executable_buffer_allocator.h",
+        "gpu_executable_va_remap_allocator.h",
+    ],
     deps = [
         ":buffer_allocations",
         ":gpu_constants",
-        ":gpu_executable_run_options",
         "//xla:shape_util",
         "//xla:util",
         "//xla:xla_proto_cc",
-        "//xla/backends/gpu/collectives:gpu_collectives",
         "//xla/backends/gpu/runtime:command_buffer_thunk",
         "//xla/backends/gpu/runtime:thunk",
         "//xla/backends/gpu/runtime:thunk_executor",
-        "//xla/core/collectives",
-        "//xla/core/collectives:collectives_registry",
         "//xla/service:buffer_assignment",
         "//xla/service:logical_buffer",
         "//xla/service:service_executable_run_options",
@@ -721,7 +723,6 @@ cc_library(
         "//xla/stream_executor:stream",
         "//xla/stream_executor:stream_executor_h",
         "//xla/tsl/platform:status_macros",
-        "@com_google_absl//absl/base",
         "@com_google_absl//absl/base:core_headers",
         "@com_google_absl//absl/container:btree",
         "@com_google_absl//absl/container:flat_hash_map",
